pyproject.toml
[project]
name = "korbpuls"
version = "0.1.0"
description = "korbpuls.de — FastAPI web application"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "Malvavisc0", email = "207609879+malvavisc0@users.noreply.github.com" },
]

dependencies = [
    "fastapi",
    "pydantic",
    "uvicorn[standard]",
    "korb",
    "llama-index",
    "llama-index-llms-openai-like",
    "jinja2",
    "python-multipart",
    "python-dotenv"
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "mypy",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/korbpuls"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E", "F", "I", "UP", "B", "SIM", "RUF",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
strict = true

warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_reexport = true
disallow_any_generics = true

files = ["src"]

exclude = [
    "^venv$",
    "^.venv$",
    "^__pycache__$",
    "^build$",
    "^dist$",
]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

# Per-module overrides — relax strict typing for tests
[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
check_untyped_defs = true
